[New]: Refactory code, Feedback alerts in the delete modal.

+++ src/components/Buttons/DeleteBtn/Modal.jsx
@@ -1,14 +1,28 @@
-import { useContext } from 'react'
+import { useState, useContext } from 'react'
 import { deletePassword } from '../../../lib/deletePassword'
-import { SavedContext } from '../../../contexts/saved-context'
-import { IndexPasswordsContext } from '../../../contexts/indexPasswords'
+import { SavedContext, IndexPasswordsContext } from '../../../contexts'
+
+const CONFIRMATION_TEXT = 'Confirmar';
+
 export const Modal = () => {
     const { setSaved } = useContext(SavedContext);
     const { indexPasswords } = useContext(IndexPasswordsContext);
 
+    const [textInput, setTextInput] = useState('');
+    const [alertMessage, setAlertMessage] = useState('');
+    const [stateAlert, setStateAlert] = useState(false);
+
     const handleDelete = (passedIndex) => {
-        console.log("CLICK", indexPasswords);
+        if (textInput !== CONFIRMATION_TEXT) {
+            setAlertMessage(`Error: ingresa el texto correcto (${CONFIRMATION_TEXT}).`);
+            setStateAlert(false);
+            return
+        }
+
         deletePassword(passedIndex);
+        setAlertMessage('Eliminado correctamente 👍');
+        setStateAlert(true);
+
         setSaved(true)
         setTimeout(() => {
             setSaved(false);
@@ -21,17 +35,17 @@ export const Modal = () => {
                     <div className="modal-body">
                         <div className="m-3">
                             <p className="fw-light text-danger">Ingresa <span className="fw-bold">&quot;Confirmar&quot;</span> para eliminar la contraseña</p>
-                            <input type="text" className="form-control bg-light text-danger border-danger" id="exampleFormControlInput1" placeholder="Google" />
-                            {/* {alertMessage &&
+                            <input value={textInput} type="text" className="form-control bg-light text-danger border-danger" id="exampleFormControlInput1" placeholder="Google" onChange={(e) => setTextInput(e.target.value)} />
+                            {alertMessage &&
                                 <div className={`alert ${stateAlert ? 'alert-success' : 'alert-danger'} w-100 mx-auto mt-4 mb-0`} role="alert">
                                     <p className="mb-0">{alertMessage}</p>
                                 </div>
-                            } */}
+                            }
                         </div>
                     </div>
                     <div className="modal-footer border-top border-danger">
-                        <button type="button" className="btn btn-secondary" data-bs-dismiss="modal" >Cerrar</button>
-                        <button type="button" className="btn btn-danger" onClick={() => handleDelete(indexPasswords)}>Eliminar</button>
+                        <button type="button" className="btn btn-secondary" data-bs-dismiss="modal" onClick={() => { setAlertMessage(''); setTextInput('') }}>Cerrar</button>
+                        <button type="button" className="btn btn-danger" onClick={() => { handleDelete(indexPasswords); setTextInput('') }}>Eliminar</button>
                     </div>
                 </div>
             </div>
